add tmux resurrect
- fix up config

+++ install.sh
@@ -4,7 +4,7 @@ export DOTFILES_DIR
 DOTFILES_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
 for file in "$DOTFILES_DIR/config/.*"; do
-  ln -sfv $file ~
+  [ -f "$file" ] && ln -sfv $file ~
 done
 
 . "$DOTFILES_DIR/setup/brew.sh"
